requirements.txt
langchain>=0.0.335
langchain-openai>=0.0.2
composio-langchain>=0.0.1
openai>=0.27.0
python-dotenv>=1.0.0
requests>=2.28.0
flask>=2.0.0
flask-cors>=3.0.10
